Update_Fixed Camera and Push/Drag
Delete PlayerStart and drag BP_PlayerBase to scene.

Check the BP_Player in OutLiner(right on the top).

Look in detail in BP_PlayerBase and search pawn to find the AutoPossess Player and change disable to Player 0.

Camera Change

Look in detail in BP_PlayerBase and search pawn.See active camera and change to BP_Camera_SideScroller